CS270 Assignment 5
CS270 Assignment 5, Phase 2 of QR Code Scavenger Hunt. First the
repository was cloned, reinstalled gems, migrated and seeded db. Event
edit page was altered to show associated Locations with checkboxes to
edit/update Events. Finally, though not assigned, I changed the
routes.rb file to point the root to event controller and index action,
just as a simple way to set the default page as a listing of events.